feature : add new routes for updating + passes all test from folder '1-Login  & basics' Nom can update a question, retrieve qeustion either with ID or position Participatio table now exists (and can be build with /rebuild-db endpoint) Error management for non existing question (neither with id or position) don't handle position error (position val < number of question)

+++ quiz-api/quiz/quiz.py
@@ -19,6 +19,7 @@ def init_db():
     try : 
         try : 
             create_db.create_question(path = create_db.PATH)
+            create_db.create_participation(path = create_db.PATH)
             return "Ok", 200
         except Exception as e :
             print(e)
@@ -61,11 +62,14 @@ def add_questions() :
 @quiz.route('/questions/<which>', methods=['DELETE'])
 def delete_question(which) : 
     # check if the user wants to delete one question or all questions
+    print(which, type(which))
     if which == "all" :
         question_id = "all"
+        code = ("No Content",204)
     else : 
         try : 
             question_id = int(which)
+            code = ("No Content",204)
         except : 
             return "Les valeurs possibles sont 'all' ou bien un nombre",401
     
@@ -78,8 +82,13 @@ def delete_question(which) :
         jwt.decode_token(token)
         
         # once the users is authenticated + the questions he wants to delete (all or a number) is correct, we can delete question
+        if question_id != "all" : 
+            question = database.retrieve_one_question(question_id)
+            if question == None : 
+                return "Not found",404
         database.delete_question(question_id)
-        return "No Content", 204
+        print(code)
+        return code
     except jwt.JwtError as e:
         print(e)
         return 'Unauthorized', 401
@@ -88,5 +97,40 @@ def delete_question(which) :
 def get_question(question_id):
     question = database.retrieve_one_question(int(question_id))
     # question.possibleAnswers = json.loads(question.possibleAnswers)
-    print(question.possibleAnswers)
+    if question == None : 
+        return "Not Found",404
     return question.toJson(), 200
+
+@quiz.route("/questions", methods=["GET"])
+def get_questions_by_position():
+    pos_id = int(request.args.get("position"))
+    question = database.retrieve_one_question(pos_id, key="position")
+    if question == None : 
+        return "Not Found",404
+    return question.toJson(), 200
+
+@quiz.route("/questions/<question_id>", methods=["PUT"])
+def update_question(question_id) : 
+    #Check if user is logged in
+    raw_token = request.headers.get('Authorization')
+    if raw_token == None : 
+        return 'Missing authorization headers', 401
+    token = raw_token.replace("Bearer ","")
+    try : 
+        jwt.decode_token(token)
+    except jwt.JwtError as e:
+        print(e)
+        return 'Unauthorized', 401
+
+    payload = request.get_json()
+    question = database.retrieve_one_question(int(question_id))
+    if question == None :
+        return "Not Found", 404
+    # Update question, don't take in account the position handling error (position > number of questions)
+    question.title = payload['title']
+    question.image = payload["image"]
+    question.text = payload["text"]
+    question.position = payload["position"]
+    question.possibleAnswers = payload["possibleAnswers"]
+    database.update_question(question)
+    return "No Content", 204
